requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
groq==0.9.0
sentence-transformers==3.3.1
joblib==1.3.2
pandas>=2.0.2
scikit-learn==1.8.0
uvicorn
python-multipart==0.0.9
